Fixed a bug

+++ DragRect/main.c
@@ -75,7 +75,7 @@ void event_loop() {
         uint32_t current = SDL_GetTicks();
         uint32_t cost = current - begin;
         uint32_t frame = 1000 / FRAMERATE;
-        uint32_t delay = frame - cost;
+        long delay = frame - cost;
 
         if (delay > 0) {
             SDL_Delay(delay);
